chore: clean up and review points
